feat: Competition slack알림 #43
