+++ alert-service/src/main/java/com/_42195km/alertservice/infrastructure/messaging/in/AchieveEventDto.java
@@ -1,11 +1,13 @@
 package com._42195km.alertservice.infrastructure.messaging.in;
 
+import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
 
 import java.util.UUID;
 
 @Getter
+@Builder
 @ToString
 public class AchieveEventDto {
     private UUID userId;
