+++ alert-service/src/main/java/com/_42195km/alertservice/infrastructure/messaging/in/AchievementKafkaConsumer.java
@@ -17,7 +17,7 @@ public class AchievementKafkaConsumer {
 
     private final MessageService messageService;
     private final ObjectMapper objectMapper;
-    @KafkaListener(topics = "achieve-achievement", groupId = "alert-group")
+    @KafkaListener(topics = "achieve-achievement", groupId = "ahieve-group")
     public void AlertAchievement(Map<String, Object> eventMap) {
         AchieveEventDto achieveEventDto = objectMapper.convertValue(eventMap, AchieveEventDto.class);
