+++ alert-service/src/main/java/com/_42195km/alertservice/infrastructure/messaging/in/CompetitionEventDto.java
@@ -1,0 +1,14 @@
+package com._42195km.alertservice.infrastructure.messaging.in;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.UUID;
+
+@Getter
+@Builder
+public class CompetitionEventDto {
+    private UUID userId;
+    private String mediaId;
+    private String title;
+}
